update the prelude for the refactored backends
