+++ crates/pacdef/src/prelude.rs
@@ -1,11 +1,14 @@
-pub use crate::backend::actual::arch::Arch;
-pub use crate::backend::actual::apt::Debian;
-pub use crate::backend::actual::{
-    dnf::Fedora, flatpak::Flatpak, python::Python, cargo::Rust, rustup::Rustup, xbps::Xbps,
-};
-pub use crate::backend::backend_trait::{Backend, BackendInfo, Switches, Text};
+pub use crate::backend::actual::apt::{Apt, AptMakeImplicit, AptQueryInfo};
+pub use crate::backend::actual::arch::{Arch, ArchMakeImplicit, ArchQueryInfo};
+pub use crate::backend::actual::cargo::Cargo;
+pub use crate::backend::actual::dnf::{Dnf, DnfInstallOptions, DnfQueryInfo};
+pub use crate::backend::actual::flatpak::{Flatpak, FlatpakQueryInfo};
+pub use crate::backend::actual::pip::{Pip, PipQueryInfo};
+pub use crate::backend::actual::pipx::Pipx;
+pub use crate::backend::actual::rustup::Rustup;
+pub use crate::backend::actual::xbps::{Xbps, XbpsMakeImplicit};
 pub use crate::backend::todo_per_backend::ToDoPerBackend;
-pub use crate::backend::AnyBackend;
+pub use crate::backend::Backend;
 pub use crate::backend::ManagedBackend;
 pub use crate::cli::CleanPackageAction;
 pub use crate::cli::EditGroupAction;
@@ -30,7 +33,6 @@ pub use crate::config::Config;
 pub use crate::errors::Error;
 pub use crate::grouping::{
     group::{Group, Groups},
-    package::{Package, Packages},
     section::{Section, Sections},
 };
 pub use crate::path::binary_in_path;
